refactor: move pdf generation from AppBase to own service
